pyproject.toml
[project]
name = "agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bs4>=0.0.2",
    "lxml-html-clean>=0.4.4",
    "newspaper3k>=0.2.8",
    "openai-agents>=0.10.5",
    "rich>=14.3.3",
    "sentence-transformers>=5.3.0",
]
